+++ docs/CHANGELOG.md
@@ -37,6 +37,7 @@
 - fixed a brief field of view flicker when exiting photo mode during Lara's special animations, such as turning to gold (regression from 1.5)
 - fixed mesh debug spheres not rendering after switching mods (regression from 1.5)
 - fixed differing ladder/hanging behavior when Lara comes to a stop from shimmying when corner shimmying is enabled (regression from 1.9)
+- fixed the game displaying a GUI error dialog when it fails to start in headless mode
 
 **TR1**
 - changed weather to be affected by the breeze